Avoid duplicate loads on a call to eaEscapeHelper
Escape Analysis will temporarily inject fake calls to <eaEscapeHelper>
to simplify analysis of OSR induce blocks for cold block escapes and to
enable heapification of objects that are passed to methods that have
been peeked into with HCR enabled.  Arguments to <eaEscapeHelper> are
values contained in local variables that might escape at that point.  In
generating calls to <eaEscapeHelper>, EA does not check whether a local
variable has already been loaded for a call, so it is possible that EA
will generate many redundant loads for a call.  This might result in
bloated IL for the duration of multiple passes of Escape Analysis, which
might hinder computing use/def information.

This change modifies the signatures of the processAutosAndPendingPushes
and processSymbolReferences methods of TR_EscapeAnalysisTools to accept
a TR_BitVector containing the symbol reference numbers that will need to
be loaded for a call to <eaEscapeHelper> rather than using a NodeDeque
containing the loads themselves.  It also modifies the signature of the
insertFakeEscapeForLoads method to accept that TR_BitVector making it
then responsible for creating the actual load operations for the call to
<eaEscapeHelper>.  That simplifies bookkeeping needed to avoid
generating redundant loads.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>