Create prex arg info from callee symbol for array-typed parameters
If this type info is missing, validateAndPropagateArgsFromCalleeSymbol()
can give up, in which case we can fail to inline even trivial methods
due to what appears to be a type incompatibility.

In particular, this would cause the actual array access to fail to be
inlined for array getter (and probably also setter) MethodHandles.