Re-generate client UID after CRIU restore operation
With JITServer Technology, each client JVM has a unique ID that
is generated during JVM bootstrap. When checkpointing a JVM that
is supposed to be used in client mode, this client UID gets
baked into the JVM snapshot, and all JVM processes that are restored
from this snapshot will have the same UID, which confuses the server.
This comit re-generates the client UID during the snapshot restore
operation so that different clients will have different UIDs.

Issue: #17022

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>