Fix inlining when the root method is a customized LambdaForm
Each intermediate result in a LambdaForm is stored in an auto. Given the
final static field load at the beginning of a customized LambdaForm,
method handle transformer is able to determine a known object for
the result of loading final fields from the MethodHandle, and it can
e.g. refine an invokeBasic() call using one into a direct call. It can't
add the known object index to the auto that represents the receiver of
such a call, since generally an auto can't have a known object index,
but it does set the known object index directly on the node.

Until now the only way that inliner has been able to find a known object
index for an argument (other than <javaLangClassFromClass> of loadaddr)
has been from the symref, so it would fail to see that the argument is a
known object, inhibiting further inlining within the callee. With this
change, inliner will now also check to see whether there is a known
object index set on the node itself, which allows it to pick up on the
result of the analysis from method handle transformer and use it to find
more call targets within the callee.