Pass separate JVM options to cmdLineTester
The existing cmdLineTester tests take options that are supposed to be
passed into JVMs that the cmdLineTester invokes and also applies it to
the JVM running cmdLineTester.

This commit creates a new variable CMDLINETESTER_JVM_OPTIONS which is
set to -Xshareclasses:none. This decouples the options under test from
and also prevents situations where the JVM that's under test ends up
generating a bad AOT method that the JVM runing cmdLineTester could
load.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>