Fix SCC AOT code validation in checkpoint-mode
In checkpoint mode we need to downgrade the target processor
to allow for portability of not only JIT code but also AOT
code. Internally we maintain two separate target processors,
`target` for JIT compiled code and `relocatableTarget` for
AOT compiled code and SCC validation. In checkpoint mode we
downgrade `target` but not `relocatableTarget`, and we do
this after validating the SCC.

This patch fixes the problem by downgrading `relocatableTarget`
and moving the code to just before SCC validation.

Co-authored-by: Irwin D'Souza <dsouzai.gh@gmail.com>

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>