Provide JITServer implementation for CRIU specific FE queries
Two frontend queries that are specific to CRIU, `inSnapshotMode()`
and `isSnapshotModeEnabled()` need to be overridden with JITServer
specific implementation that fetches the required information from
the client.
Three new fields were added to the `ClientSessionData.VMInfo` struct
which caches information per client:
      bool _inSnapshotMode;
      bool _isSnapshotModeEnabled;
      bool _isNonPortableRestoreMode;
`_isSnapshotModeEnabled` and `_isNonPortableRestoreMode` can be cached
from the very first response of the client, because they get set when the
VM is initialized and never change their value.
`_inSnapshotMode` requires special handling: if the client runs in
portable restore mode, then this field is always true. If the client runs
in non-portable restore mode, then this field starts as true, but after a
restore, it transitions to false and it stays false until the end of the JVM.
While this field returns 'true', the server must always inquire the client
about the value of this field. After it transitions to `false`, the server
can cache its value and refrain from asking the client again.
For this purpose, a new message type was introduced, `VM_inSnapshotMode`
and the `MINOR_NUMBER` of JITServer has been incremented.

Issue: #17117

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>