Fix missing concurrent scan finish notification
As part of synchronization between concurrent scanning and continuation
mounting, the continuation mounting could be blocked by concurrent
scanning, release vmaccess and waiting for current thread mutex, but
there is timing hole between release vmaccess and enter the mutex, the
notification of the mutex(the end of concurrent scanning) could be sent
before the enter of current thread mutex, then cause the mounting thread
would wait forever(hang).

Enter the mutex earlier in order to avoid the time hole.

Signed-off-by: Lin Hu <linhu@ca.ibm.com>