Add CH Table AOT Feature Flag
The commit adds a new feature flag to ensure that there isn't a CH Table
Opts mismatch. Specifically, if CH Table opts were enabled in the compile
run, then it also needs to be enabled in the load run. However, it is ok
if CH Table opts were disabled in the compile run and enabled in the
restore run.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>