Ensure failing to inline Reference.refersTo is handled properly
If TR_J9VM::inlineNativeCall fails to inline a method, it sometimes
returns NULL and sometimes returns the original call node.  If it
returns NULL and the call node is a JNI call, the caller considers
whether any special processing of the call node is required by calling
TR::Node::processJNICall.

In the case of Reference.refersTo, the call is not inlined if the
Metronome GC policy is in effect, but inlineNativeCall was returning the
original call node.  That signalled to the caller that no special
consideration of whether the call node was a JNI call was needed, but it
was in fact needed.

Fixed this by having TR_J9VM::inlineNativeCall return NULL if it is
unable to inline Reference.refersTo.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>