Handle 1D arrays with multianewarray
The tree evaluator for multianewarray on X and Z has an optimisation for 0 length
arrays that generate inline code for allocating the array. This code assumes that
the new array will always have at least two dimensions, however the VM spec permits
multianewarray to be used for 1 dimensional arrays. This does not appear to be
used by javac, but can appear in code generated by other languages, e.g. Groovy.

This commit updates the evaluator on X to check the number of dimensions. For
arrays of dimension 2 or more the original code is used with inline allocation
where possible, and for 1 dimensional arrays the much simpler code from P is
used which simply calls the helper.