Implement the TR_CatchBlockCounter relocation
The TR_CatchBlockCounter relocation introduced in eclipse/omr#6956
is now fully implemented in the relocation runtime. The symbol
created in the EDO optimization that is associated to the catch block
counter is also identified as such, and so relo records will be created
automatically for that symbol. Certain manual loads of the catch block
counter emitted during code gen are also properly relocated now.

Signed-off-by: Christian Despres <despresc@ibm.com>