Handle code cache alloction for low memory
Provide user with an option to handle the percentage of available memory to be used for code cache if the available free memory is less than the default values set at startup.
The default value of Max RAM percenetage allocated to code cache segment is set to 25% so if the available memory is 200MB then we allocate only 25%.
The order of how the value is populated for code cache will be:

Set to defaults in onLoadInternal (ex 256MB for 64 bit).
Inside preProcessCodeCacheIncreaseTotalSize calculate the free memory. If less than defaults then assign only the percentage specified.
If the user has specifed a value for code cache total then that takes precedence .

Closes: ##17011
Signed-off-by: Sajina Kandy <sajina.kandy@ibm.com>