Improve MethodHandle direct dispatch J2I-prevention transformations
...by avoiding breaking the block containing the original call. There
isn't much opportunity to optimize the inline logic anyway, so it's
better to keep the IL neater.

On code generators that support doing so (initially just x86-64), calls
to linkToStatic, linkToSpecial, and invokeBasic that have not been
refined are now changed in-place into calls to a new JIT non-helper
<jitDispatchJ9Method>, which takes the callee J9Method as an additional
(first) argument. It has JIT private linkage, but the extra argument is
treated specially so that the other arguments are passed in exactly the
way expected by (any JIT-compiled body for) the actual callee. The code
generator generates a sequence that determines whether the callee is
compiled and jumps to the JIT entry point if it is. Otherwise, it jumps
out of line to do a J2I transition.

If the environment variable TR_stressJitDispatchJ9MethodJ2I is set, then
the back end generates an unconditional jump to the J2I path, exercising
it even when the callee is compiled.