Insert branch around re performing store for awrtbar
In case of balanced GC policy, we generated code for awrtbar where in
case of Non-NULL object stored into array object, it will re performing
storing to avoid branching around which may have been cheaper. This re
performing storing in the same object field twice for awrtbarrier
somehow causing issue in unit test for PingPong using virtual threads
where we end up with incorrect state in the array of ForkJoinTask
causing one of the producer or consumer thread to be in waiting state.
Change in this commit inserts the branch around the second store where
we would jump to in case on runtime we encounter the null object store
which can bypass costly write barrier checks.

Fixes: https://github.com/eclipse-openj9/openj9/issues/17255

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>