Fix incorrect exception info in BNDCHK evaluator
When the evaluator for bound check on arrays generates code
to decompress pointers, it incorrectly sets the implicit
exception point to be the load of the pointer decompression.
This causes the garbage to collector to crash because it uses
the wrong GC atlas.

This commit correctly handles this case by setting
the implicit exception point on the cmp [mem] [reg]
which actually loads the length field of an array.

Fixes: #17171

Signed-off-by: James You <james.you@protonmail.com>