Apply less aggressive adjustment factor for non-constant args
When evaluating callees during ECS that could have an weight
adjustment factor applied, it is necessary to apply different
levels of weight reduction based on the how much benefit we can
expect from encouraging inlining of the callees being evaluated.
For example, there is more benefit to inlining constant strings,
but not always as much for non-constant strings even though the
latter case is quite common. In such a case, the more aggressive
adjustment factor is only applied to the constant string case.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>