AArch64: Add classNode argument to `genSuperClassTest`
Previously, `genSuperClassTest` checked whether the opcode of the class node
representing a class is not loadaddr if a castClass is unknown
at compile time and assumed that the class node was the second child of
the node, which served as the first argument to the function.
However, this assumption proved incorrect when the function was called
from `VMarrayStoreCHKEvaluator`.

To address this issue, this commit modifies `genSuperClassTest` to accept
a class node argument and perform the opcode check on that node instead.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>