GC CRIU: Reinit HeapRegionDescriptorExtensions (Region Obj Lists)
HeapRegionDescriptorExtensions contain multiple array of obj lists which
are inited at startup based on the startup thread count. These must be
reinitialized during restore (CRIU) according to the new GC thread
count. Specifically, when the thread count is increased, it is necessary
to divide the region object lists. This is required to improve
parallelism with GC threads and ultimately improve GC performance for
the restore environment. For background, see
https://github.com/eclipse/omr/issues/6888 (Compensate for Thread Count
Change).

Signed-off-by: Salman Rana <salman.rana@ibm.com>