Add new flag to threads to delay halt for CRIU and redo JDWP thread check
The new flag will determine if threads such as JDWP debug threads will be
halted later in the checkpoint process, after hooks, and earlier in the restore
process, before restore hooks.

JDWP threads will have this flag set on them as they need to be running during
checkpoint and restore hooks. This is done when the J9VMThread is created for
a Java thread object in jvmtiRunAgentThread. If the thread object reference
matches a JDWP thread object reference in the list the VM keeps, then the delay
halt flag is added to the new J9VMThread being created.

JVMTI extension functions to add and remove Java debug threads from the list are
added so that the updating can be done from the JDWP library.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>