Fix illegal lookupClass error caused by MethodHandles.Lookup method
Update MethodHandles.Lookup to IMPL_LOOKUP.in

If a privileged class object is passed to MethodHandles.Lookup,
it will throw an Illegal Argument Exception.
To address this issue and bypass the privileged method, using
IMPL_LOOKUP instead of MethodHandles.Lookup.

The IMPL_LOOKUP is a package-protected version of Lookup that
skips checkUnprivilegedLookupClass.

Fixes: #14544

Signed-off-by: Dipak Bagadiya <dipak.bagadiya@ibm.com>