Remove unnecessary assertion at OSCachesysv.cpp:1211
It is possible that the semaphore is removed from the system by another
process and the JVM could reach this assertion. On z/OS, this assertion
leads the JVM to exit on core dumps, which is undesirable. The design
is to ignore the startup error of default SCC and continue without using
the SCC. The assertion message is:

j9shr.2038          Exit       <OSCache enterMutex failed - j9shsem_wait
failed with error = -328433
j9shr.1009   *   ** ASSERTION FAILED **
at /openj9/runtime/shared_common/OSCachesysv.cpp:1211: ((0 ))

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>