[FFI/Jtreg_JDK21] Validate the alignment of the pointer address
The changes validate the pointer address against
the alignment value of its target layout introduced
in JDK21 if exists to ensure the pointer segment
with the valid alignment & size is passed-in or
returned during the downcall/upcall.

Note:
The changes aim to unify the validation methods in
java shared by downcall & upcall with MH filters,
in which case there is no need to do call-ins to
capture the exceptions/invalid cases in the dispatcher.
Thus, it is easy to extend the validation in java
without modifying the code in the dispatcher. 

Fixes: #17686

Signed-off-by: ChengJin01 <jincheng@ca.ibm.com>