Refactor changes from #17322 to resolve perf issues
This change addresses a 1-2% perf regression with -Xint on
Linux IA32 and Windows AMD64 platforms for Java 8.

JIT to interpreter transition (j2iTransition) should not happen
with -Xint since the JIT is disabled.

j2iTransition code changes in #17322 have unexpected compiler
side-effects which leads to a perf regression.

The work-around is to wrap the code changes from #17322 with the
ifdef for OpenJDK MethodHandles (OJDK MHs) since these changes are
used IFF OJDK MHs are enabled.

Related: #17322

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>