Load criu instead of j9 library in init check
Previously when we were relying on the dynamic linker to load libcriu.
Loading libj9criu29 was sufficient to ensure that libcriu existed.
However, since we switched to explicit loading libcriu is onl loaded at
the time of checkpoint, which means querying if criu support is enabled
is not sufficient to determine if the system can actually take a
checkpoint.

This PR loads CRIU earlier so that the user has more info before the
checkpoint is attempted.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>