Initialize JITServer SSL context post CRIU restore
When encryption is enabled for a JITServer client post CRIU restore, the
client needs to update its SSL configuration with the newly provided certificate.
This is done in ClientStream::static_init() function which needs to be called
during processOptionsPostRestore().

Moreover, if SSL encryption was specified pre-checkpoint, the hook that gets
executed, `TR::CompilationInfo::prepareForCheckpoint()`, will delete the
existing SSL context and any existing certificate.

Issue: #17786

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>