jcl: ensure JNI cache is initialized only once
initIDCache is currently subjected for race condition,
and if concurrently entered by multiple threads can
result in reference leaks. Fix this by introducing a
wrapper class that acts as a bridge, invoking the method
from the static initializer of the wrapper class, so that
the synchronization is taken care through class initializer.