[JDK20/21] Add support for JVMTI Local Variable functions
Since JDK20, all JVMTI Local Variable functions are expected to return
JVMTI_ERROR_THREAD_NOT_SUSPENDED if the thread was not suspended and
was not the current thread.

Also, there is no need to halt and resume the thread for inspection
since these functions expect the thread to be suspended.

Related: #17711

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>