Do not cache nesthost of hidden nestmate classes if extension enabled
Hidden nest member is not in the nestmember attribute of the nest host.
If extensions are enabled and the nest host is redefined, the nest host
of hidden nest member is not updated. Do not cache nest host into
class->nestHost in this case so that the nest host will always be found
in the class table for hidden nest members.

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>