Fix SVM Assert caused by missing class validation
The code in J9Inliner.cpp that implemented certain heuristics to choose
a class to inline based on profiling missed adding some SVM
validations. This commit fixes this by wrapping the code that finds a
candidate class in a heuristic region, and then adding the necessary
profiled class and instanceof validations.

These heuristics are also disabled for non-SVM AOT. Normally this code
is implicitly disabled as it only runs post-startup; however, it is
possible for the VM to transition from startup to notstartup in the
middle of a compilation, which could result in this code running in a
non-SVM compilation which is not capable of generating the necessary
validations.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>