Simplify INL frame build in System.arraycopy
The arraycopy INL builds a frame for many error cases and the value type
case which may require allocation. Simplifying the code so the frame is
always built early greatly simplifies the code and corrects a problem
introduced in #17732 (INL frame build, tear down and rebuild is no
longer allowed in general and I missed a case here). It also appear to
fix the reported issue, though I'm not entirely sure why.

Also fix a missing frame build in the Thread interrupted INL.

Fixes: #17875

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>