Add SCC layer creation for JITServer AOT cache
If a persistent shared class cache is started under certain conditions
(in a container, readonly, JITServer AOT cache enabled, JIT and AOT
compilations enabled), the SCC startup will now ignore the readonly flag
and instead initialize a temporary new (writable) top layer for the
JITServer AOT relocation mechanism to use.

New options -XX:+JITServerAOTCacheUseTemporaryLayer and
-XX:+JITServerAOTCacheUseTemporaryLayer control this mechanism; adding
the positive option will cause the layer to be created even if we are
not running in a container or readonly (useful for testing purposes) but
not in more scenarios than that, and adding the negative option will
disable the feature completely.

When using the JITServer AOT cache, it is now the default not to store
received method bodies in the local shared class cache. Setting the
environment variable TR_enableRemoteAOTMethodStorage will restore the
old behaviour. Due to these changes, the
TR_disableRemoteAOTMethodStorage environment variable, which
formerly disabled the storage of remote AOT bodies in the local SCC, is
redundant and is no longer recognized.

Signed-off-by: Christian Despres <despresc@ibm.com>