Avoid shifting null constants in compression sequences
This ensures that code generators will not generate an unnecessary shift
instruction, even if they make no effort to detect compression sequences
for null constants.

It's still always possible to traverse through the sequence to find the
original (null) reference.

Additionally, propagate the null/non-null flags from the reference node
to all of the integer-typed nodes in the sequence rather than only the
truncation (l2i), and also flag the long value immediately prior to
truncation as having zero in its high half. These properties hold, so
the flags might as well be set.