Remove incorrect TR_ASSERTs related to the reloFlags
PR 17621 adjusted where the relo flags (not to be confused with the
flags described by RELOCATION_CROSS_PLATFORM_FLAGS_MASK). The
TR_ASSERT_FATAL in TR_RelocationRecord::setReloFlags was updated
accordingly, but all the TR_ASSERTs sprinkled in the code that
initializes the header of the binary relo record was missed.

This commit removes these TR_ASSERTs because in addition to be
incorrect, they are also redundant; TR_RelocationRecord::setReloFlags
already has a TR_ASSERT_FATAL.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>