Ensure flags stored to raw relo record is 16 bits
The _flags field of the TR_RelocationRecordBinaryTemplate struct is 16
bits wide. This was to ensure that the relocation infrastructure could
store more than 4 bits of Relocation Flags (the other 4 bits are
reserved for the Cross Platform Flags).

The code in TR_RelocationRecord::setReloFlags did some validation to
ensure that the bits did not overlap; however, it did so using a uint8_t
which could lose some bits due to a shift. This commit fixes this by
ensuring a uint16_t is used.

This commit also changes how the flags are set; existing flags in the
raw relocation record are now preserved. Prior to this change, setFlags
preserved the Cross Platform Flags but setReloFlags did not preserve the
Relo Flags. Nothing currently depends on this behaviour. Therefore, to
strengthen the conceptual integrity of these APIs, this commit makes the
behaviour of both setFlags and setReloFlags consistent.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>