Update the assert to allow TR_UnloadedClassPicSite
It is possible that an entry of `TR_UnloadedClassPicSite`
(kind = `RuntimeAssumptionOnClassUnload`) remains if the
`TR_UnloadedClassPicSite` references a J9Method in the
pre-prolog. The J9Method in the pre-prolog is for the helper
to know what method to execute for a revert to the interpreter.
This address needs to be patched on a redefinition.

Fixes #17734

Co-authored-by: Kevin Langman <langman@ca.ibm.com>
Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>