Fix deadlock that occurs while suspending threads through JVMTI
getVMThread can cause a deadlock for a virutal thread if the thread is
already suspended. The suspended virtual thread halts during the mount
or unmount phase while the virtualThreadInspectorCount is set to -1.
Meanwhile, acquireVThreadInspector waits indefinitely for the
virtualThreadInspectorCount to be set to 0 in the mount or unmount
phase. To prevent the deadlock, jvmtiSuspendThread should return an
error for already suspended threads without invoking
getVMThread -> acquireVThreadInspector. This approach an be taken for
both platform and virtual threads.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>