Add stronger guarantees to System.gc()
The JCL code expects that when System.GC() is called a GC occurs and
finalizable objects are finalized. Historically, J9 has never provided
this guarantee. We have often had to modify tests to ensure a GC is
triggered at the appropriate time. Increasingly, we are finding that
more and more tests rely on this behaviour, and most importantly, core
JCL code now relies on this behaviour. This PR ensures that a call to
System.gc() triggers a GC and runs finalization.

Signed-off-by: tajila <atobia@ca.ibm.com>